requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
